Fix channel pending message count

+++ server/core/model/src/main/kotlin/io/github/vinicreis/pubsub/server/core/model/data/Channel.kt
@@ -7,7 +7,7 @@ data class Channel(
     val code: String,
     val name: String = code,
     val type: Type,
-    var pendingMessagesCount: Int = 0,
+    var pendingMessagesCount: Long = 0,
 ) {
     enum class Type {
         SIMPLE,
